Fix Windows compiler warning in tokenize.c (GH-8359)
Fix the following warning on Windows:

parser\tokenizer.c(1297): warning C4244: 'function': conversion from
'__int64' to 'int', possible loss of data.

+++ Parser/tokenizer.c
@@ -1294,7 +1294,7 @@ syntaxerror(struct tok_state *tok, const char *format, ...)
     va_end(vargs);
     PyErr_SyntaxLocationObject(tok->filename,
                                tok->lineno,
-                               tok->cur - tok->line_start);
+                               (int)(tok->cur - tok->line_start));
     tok->done = E_ERROR;
 #else
     tok->done = E_TOKEN;
